% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqtest.R
\name{chi.square.test}
\alias{chi.square.test}
\title{chi.square.test}
\usage{
chi.square.test(x, p = rep(1 / length(x), length(x)))
}
\arguments{
\item{x}{a numeric vector or matrix.}

\item{p}{a vector of probabilities of the same length as x, only used when x is a vector.}
}
\value{
\item{method}{a string describing the method used.}

\item{data.name}{a string give the name of the data used.}

\item{stat}{the value of the chi-squared test statistic.}

\item{para}{the degree of freedom of the test statistic.}

\item{pvalue}{the p-value of the test statistic.}
}
\description{
\code{chi.square.test} performs chi-squared goodness-of-fit tests on a vector and contingency table tests on a matrix.
}
\examples{
# Chi-squared goodness-of-fit tests on a vector
x <- c(12, 5, 7, 9, 8, 1)
chi.square.test(x)

# Chi-squared contigency table tests on a matrix
x <- as.matrix(c(12, 5, 7, 9, 8, 1), 2, 3)
chi.square.test(x)

}
