% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishertest.R
\name{fisher.exact.test}
\alias{fisher.exact.test}
\title{fisher.exact.test}
\usage{
fisher.exact.test <- function(x, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{a numeric vector or matrix.}

\item{alternative}{a string specifying the alternative hypothesis.}
}
\value{
\item{method}{a string describing the method used.}

\item{data.name}{a string give the name of the data used.}

\item{pvalue}{the p-value of the test statistic.}

\item{alter}{the string of alternative hypothesis.}
}
\description{
\code{fisher.exact.test} performs Fisher's exact test on 2\eqn{\times}2 contingency table.
}
\examples{
# Fisher Exact Test

x <- matrix(c(8, 7, 4, 9), 2, 2)
fisher.exact.test(x)

}
