% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemartest.R
\name{McNemar.test}
\alias{McNemar.test}
\title{McNemar.test}
\usage{
McNemar.test <- function(x, correct = TRUE)
}
\arguments{
\item{x}{a numeric vector or matrix.}

\item{correct}{a logical indicating whether using continuity correction.}
}
\value{
\item{method}{a string describing the method used.}

\item{data.name}{a string give the name of the data used.}

\item{stat}{the value of the chi-square test statistic.}

\item{para}{the degree of freedom of the test statistic.}

\item{pvalue}{the p-value of the test statistic.}
}
\description{
\code{McNemar.test} performs McNemar's chi-squared test for 2\eqn{\times}2 table.
}
\examples{

# McNemar test
x <- matrix(c(794, 86, 150, 570), 2, 2)
McNemar.test(x)

}
