% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam.R
\name{print.nonparam}
\alias{print.nonparam}
\title{print.nonparam}
\usage{
print.nonparam(x)
}
\arguments{
\item{x}{an object of class "\code{nonparam}"}
}
\value{
the argument \code{x}, invisibly
}
\description{
\code{print.nonparam} prints objects of class "\code{nonparam}" used in this package.
}
\examples{
t <- chi.square.test(1:6)
print(t)

}
