% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxontest.R
\name{wilcoxon.test}
\alias{wilcoxon.test}
\title{wilcoxon.test}
\usage{
wilcoxon.test <- function(x, y = NULL, alternative = c("two.sided", "less", "greater"),
                      mu = 0, paired = FALSE, exact = NULL, correct = TRUE)
}
\arguments{
\item{x}{a numeric vector or matrix.}

\item{y}{an optional numeric vector, of same length of \code{x} if \code{paired=TRUE}.}

\item{alternative}{a string specifying the alternative hypothesis.}

\item{mu}{an optional parameter of the null hypothesis with default \code{mu=0}.}

\item{paired}{a logical indicating this is a paired test.}

\item{exact}{a logical indicating whether using a exact p-value.}

\item{correct}{a logical indicating whether using continuity correction.}
}
\value{
\item{method}{a string describing the method used.}

\item{data.name}{a string give the name of the data used.}

\item{stat}{the value of the chi-squared test statistic.}

\item{pvalue}{the p-value of the test statistic.}

\item{alter}{the string of alternative hypothesis.}
}
\description{
\code{wilcoxon.test} performs one and two-sample Wilcoxon tests on vectors of data.
}
\examples{
x <- c(1.83,  0.50,  1.62,  2.48, 1.68, 1.88, 1.55, 3.06, 1.30)
y <- c(0.878, 0.647, 0.598, 2.05, 1.06, 1.29, 1.06, 3.14, 1.29)

# Paired two-sample Wilcoxon test.
wilcoxon.test(x, y, paired = TRUE)

# One-sample Wilcoxon test (equivalent to the above test)
wilcoxon.test(x - y, alternative = "less")

# One-sample Wilcoxon test with a position parameter.
wilcoxon.test(c(1:100, 1:100), mu=50)

# Two-sample Wilcoxon test.
wilcoxon.test(x, y)

}
